Fix jump to tests for multi-workspace projects
When jumping to a test in a multi-workspace project, the extension
currently can't find the test/implementation file. This is because
asRelativePath includes the relative path from project root for
multi-workspace settings.

See:
https://code.visualstudio.com/api/references/vscode-api

asRelativePath(pathOrUri: string | Uri, includeWorkspaceFolder?: boolean): string

When true and when the given path is contained inside a workspace folder
the name of the workspace is prepended. Defaults to true when there are
multiple workspace folders and false otherwise.

Setting includeWorkspaceFolder to false does not change anything for
single-workspace projects (as it defaults to false anyway), but it fixes
the bug for multi-workspace projects.